+++ api/__init__.py
@@ -1,6 +1,4 @@
 from functools import partial
-
-from api.backends import GuardJWTAuthBackend
 from resource import routes
 
 import falcon
@@ -11,11 +9,7 @@ from falcon_elastic_apm import ElasticApmMiddleware
 from falcon_require_https import RequireHTTPS
 from swagger_ui import falcon_api_doc
 
-from api.error_handler import (
-    BadRequestHandler,
-    InternalServerErrorHandler,
-    UnsupportedMediaTypeHandler,
-)
+from api.error_handler import BadRequestHandler, InternalServerErrorHandler, UnsupportedMediaTypeHandler
 from api.media_handler import XMLHandler, YAMLHandler
 from api.middleware import NegotiationMiddleware
 from api.spec import Spec
@@ -29,31 +23,10 @@ def api(title, version):
 
     middlewares = [NegotiationMiddleware()]
 
-    def user_loader(token):
-        return {"user": token}
-
     if ArgReader.db.auth:
         _extracted_from_api_9(log, middlewares)
     else:
         log.notice("JWT authentication disabled")
-
-    if ArgReader.db.oauth2:
-        log.notice("OAuth2/JWT authentication enabled")
-
-        guard_jwt_auth_backend = GuardJWTAuthBackend(
-            client_id=ArgReader.db.oauth2_client_id,
-            client_secret=ArgReader.db.oauth2_client_secret,
-            user_loader=user_loader,
-            token_check_uri=ArgReader.db.oauth2_token_check_uri,
-            verify=ArgReader.db.oauth2_verify,
-            auth_header_prefix=ArgReader.db.auth_header_prefix
-        )
-        exempt_routes = ["/api/doc", "/api/doc/swagger.json", "/dev/token"]
-        middlewares.append(
-            FalconAuthMiddleware(guard_jwt_auth_backend, exempt_routes)
-        )
-    else:
-        log.notice("OAuth2/JWT authentication disabled")
 
     if ArgReader.db.https:
         log.notice("Force to use HTTPS instead of HTTP")