+++ scripts/start.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# Copyright (c) 2020-2029 GUARD Project <guard-project.eu>
+
+python main.py